refactor: 해당 repository 일관성 있게 리팩토링
