+++ src/main/java/com/project/trainingdiary/repository/schedule/ScheduleRepository.java
@@ -10,9 +10,15 @@ import org.springframework.data.jpa.repository.Query;
 public interface ScheduleRepository extends JpaRepository<ScheduleEntity, Long>,
     ScheduleRepositoryCustom {
 
-  @Query("select s.startAt from schedule s where s.startAt >= ?1 AND s.startAt <= ?2")
+  @Query("select s.startAt "
+      + "from schedule s "
+      + "where s.startAt >= ?1 "
+      + "and s.startAt <= ?2")
   Set<LocalDateTime> findScheduleDatesByDates(LocalDateTime startAt1, LocalDateTime startAt2);
 
-  @Query("select s from schedule s where s.startAt >= ?1 AND s.startAt <= ?2")
+  @Query("select s "
+      + "from schedule s "
+      + "where s.startAt >= ?1 "
+      + "and s.startAt <= ?2")
   List<ScheduleEntity> findByDates(LocalDateTime startAt1, LocalDateTime startAt2);
 }
